Fix architecture review findings: date bugs, Decimal128, securities, pagination, ops
Correctness (verified by regression tests in tests/test_fixes.py):
- Date parameters now match the whole calendar day and ranges are inclusive
  of both end days (get_settlements returned 0 forever; to_date trades were
  silently dropped). New day_range/date_window helpers in services/_common.
- Malformed dates return {"code": "INVALID_DATE"} envelopes (never raise);
  REST maps them to HTTP 400.

Contract changes (all transports + both evaluation twins in lockstep):
- count in list envelopes is now the TOTAL matching documents (breaking).
- Monetary/quantity/rate fields migrated to Decimal128 in Mongo, Decimal in
  the Pydantic models, exact strings on the wire via a Decimal GraphQL
  scalar in all three GraphQL layers.
- Timestamps serialize with explicit UTC offset (+00:00).
- New securities surface: get_security/list_securities across service, REST
  (/securities), Ariadne, Strawberry, Graphene, and MCP (15 -> 17 functions).

MCP completeness:
- skip parameter on all list tools; Optional type hints fixed.
- Startup lifespan actually registered with FastMCP (was dead code).
- tests/test_mcp.py adds the missing MCP == service parity leg (17 tools).

Operations:
- /ready endpoints ping MongoDB on all four apps; K8s readiness probes moved
  to /ready; namespace.yaml added; imagePullPolicy -> IfNotPresent.
- MONGO_ERROR envelopes log the stack server-side and return a generic
  message (no internal detail leak). Single-level structured JSON logs.
- GitHub Actions CI: Mongo service container + seed + full suite.

94 tests passing (was 82). Docs updated: ARCHITECTURE, AGENTS, README.

+++ src/bank_ods/graphql_graphene/types.py
@@ -92,7 +92,7 @@ class TransactionSummaryItem(graphene.ObjectType):
     transactionType = graphene.String(required=True)
     status = graphene.String(required=True)
     count = graphene.Int(required=True)
-    totalNetAmount = graphene.Float(required=True)
+    totalNetAmount = graphene.Decimal(required=True)
 
 
 TransactionSummaryList = _list_wrapper("TransactionSummaryList", TransactionSummaryItem)
@@ -102,7 +102,7 @@ class ProjectedBalance(graphene.ObjectType):
     accountId = graphene.String(required=True)
     currency = graphene.String(required=True)
     asOfDate = graphene.String(required=True)
-    closingBalance = graphene.Float()
-    pendingCredits = graphene.Float()
-    pendingDebits = graphene.Float()
-    projectedBalance = graphene.Float()
+    closingBalance = graphene.Decimal()
+    pendingCredits = graphene.Decimal()
+    pendingDebits = graphene.Decimal()
+    projectedBalance = graphene.Decimal()
